Fixed lock to use pattern prescribed by redis
RedisLock did not clear locks from redis when they are timed out, so a
client failing to release a lock created a permanent deadlock (the lock
would throw a TimeoutException if one was supplied but that's not
particularly helpful if you can't clear the remote lock).

I've implemented the design pattern described at
http://redis.io/commands/setnx in order for locks to properly flush and
included a unit test to verify.