Added Tests showing that equal items are not removed
If the custom type has an overloaded Equals Operator the item is not always removed
since the redis implementation is serializing the object to compare if the objects
are not exactly identical the item will not be removed from the list. However
it will be removed from a List<T> implementation.

I feel that this should be changed to be identical so that the IRedisList<T> operates
with the same behavior as List<T> since in every other case it does.