Changed the ExecCachedLua's cache to an instance member and updated its usage to ConcurrentDictionary's higher level API.
Changed the ExecCachedLua's cache to an instance member and updated its
usage to ConcurrentDictionary's higher level API.

I'll explain:
-------------
1. If the RedisClient is used across multiple Redis machines, and the
same LUA script is used in both - After the LUA call-and-cache to the
first Redis instance, trying to execute the LUA on the second instance
will result in an exception ("NOSCRIPT"), re-caching the script and
executing it again.
This scenario can be eliminated by changing the cachedLuaSha1Map from a
static member to an instance member.

2. I've used the ConcurrentDictionary.GetOrAdd instead of writing its
implementation ourselves.

3. In the ExecCachedLua's catch, I've changed the ConcurrentDictionary's
access to AddOrUpdate, which has an internal white(true) loop to try to
set the script in any case of failures in the indexer's TryAddInternal.