Respect ConnectTimeout when in netstandard2
We were doing some testing of ServiceStack.Redis and noticed that the
connect timeout is not working and would block when running on
dotnetcore. I believe the conditional that was there was needed when we
the project was targetting netstandard1.3 but now with 2.0 that whole
conditional is not needed and now the timeout is being respected after
this change.